fix: align session notices with the transcript rail and split oversized web sources
Session-level notices (runtime recovery, stop attention, resume failure)
now ride the transcript's inset rail with symmetric vertical spacing
instead of running edge to edge flush against the scrolled transcript.

CI lint follow-ups: `web/src/systems/session/index.ts` and
`adapters/session-api.ts` exceeded the 500-line production budget, so the
busy-input re-exports moved to `busy-input.ts`, the history/recap/usage/
ledger reads moved to `adapters/session-history-api.ts`, and the
attachment surface is star-exported from its grouped module. The
control-plane docs page is formatted with oxfmt (the pre-commit hook only
formats `.md`).
