+++ web/src/systems/session/adapters/session-api.ts
@@ -3,20 +3,14 @@ import { apiClient, apiRequestFailed, requireResponseData } from "@/lib/api-clie
 import type {
   ApproveSessionParams,
   CreateSessionParams,
-  FetchSessionEventsParams,
-  SessionLedgerResponse,
-  SessionEventPayload,
   SessionPayload,
   SessionPromptRequest,
   SessionPromptResponse,
   SessionPromptSendResult,
-  SessionRecapPayload,
   SessionRepairPayload,
   SessionRepairQuery,
   RenameSessionRequest,
   SetSessionRuntimeRequest,
-  SessionUsagePayload,
-  TurnHistoryPayload,
   SessionStopResult,
 } from "../types";
 import {
@@ -49,6 +43,14 @@ export {
   fetchSessionInteractions,
   type FetchSessionInteractionsOptions,
 } from "./session-interactions-api";
+export {
+  fetchSessionEvents,
+  fetchSessionHistory,
+  fetchSessionLedger,
+  fetchSessionRecap,
+  fetchSessionUsage,
+  SessionLedgerUnavailableError,
+} from "./session-history-api";
 
 export type {
   ApproveSessionParams,
@@ -310,46 +312,6 @@ export async function clearSessionRuntime(
   return requireResponseData(data, response, `Failed to clear runtime for session "${id}"`).session;
 }
 
-export async function fetchSessionRecap(
-  workspaceId: string,
-  id: string,
-  limit?: number,
-  signal?: AbortSignal
-): Promise<SessionRecapPayload> {
-  const { data, error, response } = await apiClient.GET(
-    "/api/workspaces/{workspace_id}/sessions/{session_id}/recap",
-    {
-      params: {
-        path: { workspace_id: workspaceId, session_id: id },
-        query: limit === undefined ? undefined : { limit },
-      },
-      signal,
-    }
-  );
-  if (apiRequestFailed(response, error)) {
-    throwSessionRequestError(response, error, `Failed to fetch session recap "${id}"`, id);
-  }
-  return requireResponseData(data, response, `Failed to fetch session recap "${id}"`).recap;
-}
-
-export async function fetchSessionUsage(
-  workspaceId: string,
-  id: string,
-  signal?: AbortSignal
-): Promise<SessionUsagePayload> {
-  const { data, error, response } = await apiClient.GET(
-    "/api/workspaces/{workspace_id}/sessions/{session_id}/usage",
-    {
-      params: { path: { workspace_id: workspaceId, session_id: id } },
-      signal,
-    }
-  );
-  if (apiRequestFailed(response, error)) {
-    throwSessionRequestError(response, error, `Failed to fetch session usage "${id}"`, id);
-  }
-  return requireResponseData(data, response, `Failed to fetch session usage "${id}"`).usage;
-}
-
 export async function repairSession(
   workspaceId: string,
   id: string,
@@ -432,28 +394,6 @@ export async function clearSessionConversation(
   return body.session;
 }
 
-export async function fetchSessionEvents(
-  workspaceId: string,
-  id: string,
-  params?: FetchSessionEventsParams,
-  signal?: AbortSignal
-): Promise<SessionEventPayload[]> {
-  const { data, error, response } = await apiClient.GET(
-    "/api/workspaces/{workspace_id}/sessions/{session_id}/events",
-    {
-      params: {
-        path: { workspace_id: workspaceId, session_id: id },
-        query: params,
-      },
-      signal,
-    }
-  );
-  if (apiRequestFailed(response, error)) {
-    throwSessionRequestError(response, error, `Failed to fetch session events "${id}"`, id);
-  }
-  return requireResponseData(data, response, `Failed to fetch session events "${id}"`).events;
-}
-
 export async function approveSession(
   workspaceId: string,
   id: string,
@@ -471,50 +411,4 @@ export async function approveSession(
   if (apiRequestFailed(response, error)) {
     throwSessionRequestError(response, error, "Failed to approve permission", id);
   }
-}
-
-export async function fetchSessionHistory(
-  workspaceId: string,
-  id: string,
-  signal?: AbortSignal
-): Promise<TurnHistoryPayload[]> {
-  const { data, error, response } = await apiClient.GET(
-    "/api/workspaces/{workspace_id}/sessions/{session_id}/history",
-    {
-      params: { path: { workspace_id: workspaceId, session_id: id } },
-      signal,
-    }
-  );
-  if (apiRequestFailed(response, error)) {
-    throwSessionRequestError(response, error, `Failed to fetch session history "${id}"`, id);
-  }
-  return requireResponseData(data, response, `Failed to fetch session history "${id}"`).history;
-}
-
-export class SessionLedgerUnavailableError extends SessionApiError {
-  constructor(id: string) {
-    super(`Session ledger not materialized: ${id}`, 404, id);
-    this.name = "SessionLedgerUnavailableError";
-  }
-}
-
-export async function fetchSessionLedger(
-  workspaceId: string,
-  id: string,
-  signal?: AbortSignal
-): Promise<SessionLedgerResponse> {
-  const { data, error, response } = await apiClient.GET(
-    "/api/workspaces/{workspace_id}/memory/sessions/{session_id}/ledger",
-    {
-      params: { path: { workspace_id: workspaceId, session_id: id } },
-      signal,
-    }
-  );
-  if (apiRequestFailed(response, error)) {
-    if (response.status === 404) {
-      throw new SessionLedgerUnavailableError(id);
-    }
-    throwSessionRequestError(response, error, `Failed to fetch session ledger "${id}"`, id);
-  }
-  return requireResponseData(data, response, `Failed to fetch session ledger "${id}"`);
 }
